Suppress "spurious" -Wmaybe-uninitialized in SWIG
Suppress a -Wmaybe-uninitialized (which is promoted to an error) in the
SWIG MATLAB bindings. The compiler is newly detecting this because
hidden visibility allows it to inline more aggressively. The variable in
question can't actually be used uninitialized, but only because of an
unannotated noreturn function. (Due to the missing annotation, the
compiler can't prove that the dubious code is unreachable, and is thus
correct to warn.)

See also https://github.com/jaeandersson/swig/pull/73.