Tweak logic deciding if an external is enabled
Modify the logic that decides whether or not an external project is
enabled to respond to externals marked `ALWAYS` by creating a hidden
option (on by default) to enable the external, rather than enabling it
literally always. This allows users with special needs (e.g. the CI in
particular) to have a hidden mechanism to disable such externals.