Make director a CMake external
Use newly added SOURCE_SUBDIR support to point to the superbuild
CMakeLists.txt when building director. This allows us to build director
as a CMake external, rather than relying on a PODS Makefile in a
subdirectory of the source.

This also allows us to get rid of the ugly BUILD_COMMAND_DIR hack.