Include license of stuff from CMake upstream
The modules we copied from CMake upstream refer to a separate file, and
suggest to "substitute the full License text for the above reference".
Rather than do that (which would bloat all of the modules being so
distributed and cause unnecessary deltas versus being straight copies
from upstream), just stick a copy of Copyright.txt next to the files in
question.