Delay setting up options
Move the logic to set up drake's (external project) options to a macro,
rather than executing it immediately on inclusion of the file. This is
necessary for some of the dependent options to be set up properly on an
initial run (which matters especially for CI), as the variables that
control them may not be set up yet when the file is included.