"Fix" export of template instantiations
For some reason, template instantiations are not permitted to have
(C++11) attributes, which requires that the old style __attribute__ be
used instead to specify the ELF symbol visibility. Accordingly, modify
the definition of `DLLEXPORT` as used in drakeMexUtil.{cpp,h} to use the
"old" form. (We were - and still are, in other places - using C++11
attributes because they are more convenient, in particular because we
can use them with fewer compiler checks. Note that GCC does the
reasonable thing and accepts the C++11 attribute.)