Import cpplint
Copy cpplint.py from the google_styleguide external into drake proper.
Besides making maintenance easier (e.g. considering we are using a
private fork at the moment), this will eventually make it possible for
the CI cpplint build to not need to check out or actually "build"
anything besides the drake repository itself, which will both simplify
that "build" significantly, and remove the need for us to introduce
hacks into the build system to support a less-than-minimal build for the
sake of the CI cpplint "build".