Fix submodule sync dependencies
Fix setting up the dependencies on the global submodule sync target so
that all projects with an update step depend on it. The old logic was
only adding the dependency — to ALL projects — when Ninja is used. This
latter is necessary due to quirks in how CMake computes dependencies for
Ninja (fixed in 3.7), but the former, which should be done for obvious
reasons, was missing. (This got messed up setting up the dependencies
for the Ninja case.)