cmake: Stop generating the PODS include directory in the build tree
We no longer require a PODS layout in the build tree (it was broken
anyway because we were not adding the export headers to it).  Replace
use of the `pods_install_headers` function with a direct invocation of
`install(FILES)`.  This avoids re-running CMake every time a header is
changed.

Closes: #2307