Corrected inertia measures.
Use cylinders instead of meshes for some links, for processing speed.
Leave meshes for links at begin and end of arm, for model accuracy.
Changed colors to better match reality. In our arm, only a part of the link
is orange, and this colors the entire link orange, but its an improvement
to see each link.
Adding mesh files in dae format and model.config for future work in Gazebo.

Replace link 5’s mesh with a cylinder.

Update collision comments