16.04 prereqs: make distro check actually work.
Turns out that 'if' clauses can have syntax errors, and the script will just
blunder on. If they are instead spelled in the perl-ish "<test> || die" idiom,
the script will indeed fail on a syntax error in the test, printing both the
syntax error and a (potentially misleading) error message.

The distro check was using an integer comparison, rather than string (afaik
bash cannot compare floats). This effectively disabled the check, because the
syntax error scotched the entire 'if' block and the script just continued.

 * Fix the comparison to actually work: string "==" is the thing.
 * Rewrite the fail checks such that bash will notice syntax errors.