Add todo and doc to RigidBody::adjacentTo
The adjacentTo method is a bit ambiguous.  The definition of adjacency
isn't clear.  Furthermore, the method is really only used by the clique
definition.  The name feels more generic than that and might invite
use in other places which would be inappropriate for its purpose (e.g.,
if querying two arbitrary bodies simply to see if they are "connected"
in the tree.  This would return false to anything "connected" to the world
by a floating joint.)