Killed the rigid_body_collision_element
The class provided almost no novel functionality.  The single exception is
that it provided constructors through which the underlying RigidBody could
be set.  However, because the member belonged to the base class, the
constructors were added to the base class.

Otherwise, all references to the class were replaced with references
to the base class (DrakeCollision::Element).