Add and use Drake LCM types C++ library
Add an INTERFACE library on which we can hang the include directory
interface for the Drake C++ LCM type bindings. Use this also to
propagate the dependency on said bindings being generated, rather than
using add_dependencies on the custom target. Change includes of LCM type
bindings to consistently omit the 'lcmtypes/' prefix, as is done by the
generated bindings themselves. Link to imported lcmtypes_bot2-core-cpp
INTERFACE library, because a) we use it, and b) we newly need its
include directories due to removal of 'lcmtypes/' from the include
directives. (This also allows us to drop the manual include directive,
and "fixes" #3352 since the imported includes are by default SYSTEM.)