Collision clique (#3536)
* Rename "CollidesWith" with "CanCollideWith"

The method "CollidesWith" on both the collision element and rigid body
is misleading.  It suggests that a collision computation is performed.
In reality, it checks arbitrary criteria to determine if the pair of
entities should even be considered collision candidates. The new name
better reflects this reality.

* Add cliques to collision element

Added the per-collision element clique representation, a handful of methods
on that data (counting, adding, retrieving list of cliques) and some unit
tests to confirm that it works.

* Add cliques into the rigid body class

RigidBody class now has knowledge of and ability to manipulate the
collision cliques.

* Rename Model::readElement to Model::FindElement

1) The form of the method didn't match the style guid.
2) "Read" suggested work that isn't actually being done.
3) Preparation for creating the "FindMutableElement" required for bulding
   collision cliques.

* Add Model::FindMutableElement

In order to build collision cliques, entities need access to mutable
versions of the elements.  This provides that access.

* Rigid body tree builds collision cliques

As part of the tree "compilation" process, the cliques for all of the
collision elements are defined.  The "automated" clique building includes
the following:

1) One clique per body is created - all elements (even if only one) which
   belong to the body are assigned to that clique.
2) For all body *pairs* (b1, b2), if b1.CanCollideWith(b2) is false, a
   clique is created and both bodies are assigned to that clique.  This
   relies on the required symmetry of CanCollideWith.

This required moving the RigidBody.AddToCollision from private to public.

* Modify the creation of cliques to omit cliques for single-element bodies

If a RigidBody has only a single element, no clique will be created for
that element.

* Add todo and doc to RigidBody::adjacentTo

The adjacentTo method is a bit ambiguous.  The definition of adjacency
isn't clear.  Furthermore, the method is really only used by the clique
definition.  The name feels more generic than that and might invite
use in other places which would be inappropriate for its purpose (e.g.,
if querying two arbitrary bodies simply to see if they are "connected"
in the tree.  This would return false to anything "connected" to the world
by a floating joint.)

* Forgot to add collision elements to bodies

The rigidbody tree now adds the collision element to the corresponding
rigid body.  TO do so, the method, which had originally been incorrectly
made private, is now public.

* Clique removes the need for rigid body collision element collidable test

The RigidBodyCollisionElement had an implementation of CanCollideWith that
referred back to the body.  The clique already encodes that logic such
that the upward reference is no longer required.

Furthermore, the overriden version supplanted the base version and would
have ignored cliques.

* Correct cpplint issues

* Reviewer comments

Addresses cosmetic issues, predominantly in the code documentation.

* Simplify assigning collision element to rigid body from tree

The rigid body tree was using three separate methods on the body to
assign a collision element (and associated data) to the body.

The redundancy has been rolled into a single method and the old methods
have been removed.

* Make code documentation and formatting tweaks per reviewer comments

Also changed to handle self-collision in a more encapsulated manner.

* Rename number_of_cliques() to get_num_cliques()

This brings the code up to match the style guide better.

* Corrected cpplint problem

Removed stray whitespace

* Cosmetic clean up per review comments

* Clean up clique generation ownership

1. RigidBodyTree owns clique generation.
2. Clique ids are only unique at the RBTree level
3. RigidBody is given the chance to consume a clique id for self collision
   and reports if it does or not (allowing the RBTree to respond
   appropriately.)

* Tests for RigidBody::adjacentTo

* Add test coverage for RigidBodyTree::CreateCollisionCliques

The testing passes through the RigidBodyTree::compile method.  Tests the
simple cases of self-collision and "can't collide with" cliques.

* Correct cpplint issues

Some white space and line length issues resolved.

* Clean up additional white space issues

* Cosmetic tweaks per reviewer comments

* Fixed member ownership in new unit tests

THe ownership of pointers is now made clear and explicit.

* Shuffling pointer management as per comments

* Bring testing classes in line with coding style

* Fixed cpplint errors (and ran clang format)

* Fixed typo