Tweak to definition of "pose"
DrakeShapes::Element has two transforms: T_element_to_world and
T_element_to_local.  That means the shape is defined in element space, it
is moved to be aligned to the rigid body in the rigid body's "local"
space.  Finally, when the rigid body's local to world transform is
provided, the T_element_to_world matrix can be defined.

The DrakeCollision::Element description of "pose" now matches that.