Kill rb collision element (#3657)
* Add cliques to collision element

Added the per-collision element clique representation, a handful of methods
on that data (counting, adding, retrieving list of cliques) and some unit
tests to confirm that it works.

* Killed the rigid_body_collision_element

The class provided almost no novel functionality.  The single exception is
that it provided constructors through which the underlying RigidBody could
be set.  However, because the member belonged to the base class, the
constructors were added to the base class.

Otherwise, all references to the class were replaced with references
to the base class (DrakeCollision::Element).

* Cleaning up cpplint issues

* Clean up constructor

Remove dead constructor, add documentation.

* Correct cpplint issue

* Filled in missing constructor

The sdf and urdf parsers need to instantiate the collision element without
knowing the geometry details.  I needed a constructor which would allow
instantiation only with rigid body and pose.

* Add class documentation to clarify class-level vocabulary

Rather than defining terms at each function location, the class defines
the terms in aggregate and uses those terms in the subsequent
documentation.

* Tweak to definition of "pose"

DrakeShapes::Element has two transforms: T_element_to_world and
T_element_to_local.  That means the shape is defined in element space, it
is moved to be aligned to the rigid body in the rigid body's "local"
space.  Finally, when the rigid body's local to world transform is
provided, the T_element_to_world matrix can be defined.

The DrakeCollision::Element description of "pose" now matches that.

* Tweaked code docs a bit more

* More comment modifications

Tighten up the transform language a bit more.

* Update kuka demo to remove rigid body collision element

The demo has been updated in master -- this allows for a successful merge
and build.

* Fix indentation error