Extend collision detection to return collision information
The standard interface for collision detection returns bodies colliding.
A lower level representation would be the collision elements.  This is
particularly important because there is not necessarily a one-to-one
mapping between bodies and elements.  A body may be represented by several
collision elements thus, if a body is colliding, it is ambiguous which
element led to the collision.