change implementation of SKIP_DRAKE_BUILD cmake option
Instead of null-ifying the build and install commands, we remove
the drake target from 'all'.  All projects that depend on drake
are also removed from 'all'.  In order to preserve the old
behavior, a new target named drake-configure is added to 'all'
when SKIP_DRAKE_BUILD=ON.

Also changed: now director is built with drake as a dependency,
so director and signalscope build after drake.  As a result, the
visualization startup test has been disabled.  Going forward, the
dashboards should run ctest for the director external.