Improve how externals are build-skipped
Add new logic to handle `SKIP_<PROJECT>_BUILD` for any external, not
just Drake proper. When set for a project, the project and any dependees
are excluded from ALL, but the project's configure step is still
executed (unless a dependency is skipped). This way, a skipped external
can still be built by naming it explicitly as a target to build, and we
avoid the hack of monkeying with the external's build and install
commands.