Corrected a math error in computing net response
The sampled contact manifold had an error in computing the net response.
Instead of offsetting the losses to torque due to moving the force, I was
magnifying them.  This corrects that error.

The unit tests have been changed so that they use a different set of
semantics to determine validity; that either the single net wrench or the
set of wrenches should end up creating the same effect on an arbitrary
point.  Confirmed that these tests distinguishes between the error in
the code and the current state.