Reformulate ContactForce
The contact force original stored the translational force as components:
normal and tangent.  I do a lot of computations with just the normal
component, so this was efficient.  However, there are circumstances in
which the normal direction could not be extracted again.

To that end, the ContactForce is now initialized on force and normal
direction.  It is more expensive to get the components, but it is more
general.

NOTE: This did lead to a slight perturbation of the tests due to numerical
issues which led to revelation of a mean shift theorme bug.  Now fixed.