Allow lazy linking on sanitize builds
When sanitizers are in use, don't add the flags to require strict
linking (no unresolved externals). This is needed as the compiler does
not include necessary sanitizer libraries when linking libraries, and is
is difficult to impossible to specify them manually. We'll still use
strict linking for the general case, so shouldn't lose much error
detection, but this should allow the sanitize builds to work without too
many headaches.