Update minimum moment point calculations (aka center of pressure) (#3985)
* Add the definition of a single contact force

The application point and the wrench.

* Update contact _force

1. Fix const/non-const getters
2. Add .cc file including the constructors and explicit instantiation.
3. Add .cc file to CMake
4. Make it copyable/movable
5. Proper initialization

* Add preliminary resultant computation class

The .h and .cc files (with the .cc file added to cmake).  Still have a
number of todos (i.e., *all* of the functionality.)

* Initial unit test added

Simple unit test place holder for testing the resultant force calculation.

* Correct cmake listing of contact_resultant_force_calculator.cc

* Correct some cosmetic issues.

* THe resultant is a wrench, not a force

Renamed methods and members accordingly.

* Simplify "dirtiness" - it's either completely dirty or clean

The two values are best computed together and they are invalidated together,
so, the whole thing is either dirty or clean.

* Caching made mutable so computing methods can be const

The caching is supposed to be completely private.  To that end, the methods
for acquiring the values are declared const -- the data provided by the
user will remain unchanged.

* Complete interface for reporting forces

Added additional interface where a `ContactForce` can be passed directly.
Also declared the type of vector.

* Add getter for the composed force

THe ContactForce includes a single force that wraps normal and tangential
components.

* Implementation of computing resultant

The resultant computation is now fully implemented.  (Modulo errors).
It still needs to be tested.

* cpplint corrections

* Initial unit tests; contact force tests

I test the constructors and the getters.

* Update ContactResultantForceCalculator interface

1) Add a default constructor.
2) Change the order of the force components to be the same as the
   ContactForce class.

* More unit tests -- adding forces

Test two pieces of code simultaneously:

1) The interfaces to add forces to the calculator.
2) The case where only a single force is applied (that force should be
   returned as the resultant.)

* Fixed bad implementation of mean-shift theorem

I had a typo and bookkeeping error in the application of the mean shift
theorem.  No longer.

* Added center-of-pressure tests

Tested the case where the points should successfully produce the correct
center of pressure -- satisfying the requirements.

* Corrected second mean-shift issues.

* Rename ComputeCachedData

It emphasizes the fact that the data is cached, but ignores *what* the data
is.  New name addresses that.

* Handle and test a "zero pressure" case

This is where the forces being applied have a zero normal component.
The choice of application point is arbitrary.  So, the first point is
selected as application point.

* More unit tests

Test the cases of non-parallel normal forces.  Both cases where the
application points are planar and where they are not.

* Significant update on the class documentation

The class documentation has been beefed up.  Ready for PR.

* Change resultant calculator to return a contact force

Rather than invoking two different methods to get the two parts, return
a ContactForce so that the normal direction can still be extracted from
the resultant.

* Reformulate ContactForce

The contact force original stored the translational force as components:
normal and tangent.  I do a lot of computations with just the normal
component, so this was efficient.  However, there are circumstances in
which the normal direction could not be extracted again.

To that end, the ContactForce is now initialized on force and normal
direction.  It is more expensive to get the components, but it is more
general.

NOTE: This did lead to a slight perturbation of the tests due to numerical
issues which led to revelation of a mean shift theorme bug.  Now fixed.

* Drake assert contact force normal unit-lengthedness

The contact force asserts the unit length.  This revealed that in the case
where there is no translational resultant force, I was *not* creating a
ContactForce with a unit normal.  Corrected.

* Various cosmetic changes -- clang format and cpplint

* Update calculator documentation

* Eliminate now-redundant cache system on calculator

The change in the interface removes the need to cache the answer.

* I can't type "deformed"

* Eliminate default constructor

In retrospect, there was no need for the default constructor.  It
initialized the values to meaningless values that cannot be changed after
construction.

* Updated documentation as per review

* Cosmetic changes due to reviewer comments

* Introduce SpatialForce type

SpatialForce is a typdef of a column vector in R6 (like WrenchVector).

1) Add the typedef.
2) Document SpatialForce vs WrenchVector
3) Update ContactForce to return this type.

* Change the zero-translational force minimum moment point

The previous code selected the first point as the minimum moment point
when the normal forces summed up to zero.  At that point, every point is
equally good.  However, this made the output dependent on the order in
which the forces were added.

This eliminates that dependency.

* Bit more doxygen clean up

* Cleaned up ContactForce getter logic

I had a hodge podge of getter logic; there was a total disarray in logic
regarding returning references and consts.  Now, all getters are const.
It returns references to members and actual values to constructed values
(i.e., normal and tangential components.)

* Re-order contact force parameters

Previously it was: point, force, normal, torque
Now it is: point, normal, force, torque.

* Missed a cpplint issue

Two line were too long.

* Cleaned up cosmetic issues

* Removed now-unused utility methods

* Various cosmetic changes based on reviewer comments

* Unit test for central axis property

The idea is that I could pick an application point anywhere on theline and
get the same moment.  However, I'm not entirely sure the test introduced
here is overly meaningful.

* Updated as per discussion

1) Extended comments.
2) Cleaned up initialization of vectors.
3) Eliminated otherwise pointless test.

* Addressing reviewer comments

1) New language in the documentation.
2) Added a new unit test case.

* Fixed documentation typos/errors as per reviewer request.