Fix setting pkg-config search paths
Fix lying comment about adding the pkg-config path from drake's build
directory to the search paths. This would be horribly broken if we
actually relied on that, however that's not what we were actually doing.
What the logic actually does is add to the search path the location
where our externals put their .pc's. Update the comment accordingly, and
also change the way we specify the path to make this much more obvious.
Also, add the path with just 'lib' rather than the correct subdirectory,
${CMAKE_INSTALL_LIBDIR}, which may or may not be the same, so that we
can find .pc's for externals that install to the wrong directory.