Don't export drake tests to superbuild (temporary)
Temporarily remove the flag to export the tests from Drake proper to the
superbuild. This will allow us, as a stop-gap until we get subprojects
working, to submit the tests for everything else to CDash while still
keeping at least "Drake" and "everything else" separated, and without
running Drake's tests twice.

Once subprojects are working, we'll want to revert this change...