Collision filter groups (#4773)
This implements collision filter groups in C++ (see issue #4204).

The behavior in C++ resembles that of Matlab but does not reproduce it exactly.

In C++, a single urdf file can be loaded multiple times (in matlab, this would lead to an error.)
In matlab, links (aka bodies) can be added/removed to collision filter groups programmatically, outside of parsing. In this C++ implementation, only parsed collision filter groups are supported. There is a TODO indicating that this needs to change as per this issue: #4729.
Similarly, matlab provides runtime querying (e.g., given the name of the a collision filter group, can look up the bit id.) C++, not so much. It is also worth noting, that this matlab functionality must be run before it is converted into a C++ object via the mex functionality.