1. Introduced a StateFeedbackController interface, which has 2 inputs (measured and desired state), and 1 output. This extends the diagram. And is intended to be a base interface. Needs concrete implementations.
2. Introduced a ModelBasedController, which is a Controller, and own an
instance of a RBT.

3. ** Changed the PID controller's interface to Controller **. It used to
be 2 inputs(e, and edot) and 1 output y = kp * e + kd * edot + ki *
int(e), where e is q - q_d. In the past, the user has to construct a
bunch of little blocks to compute e, and then invert the y then feed it
into the RBP. Now, all you need to do is call
connect(rbp.state_output(), pid.measured_state_input).
connect(pid.output, rbp.input), and connect(your_reference_state_source,
pid.desired_state_input).
The new PID controller still depends on the old one internally, I am
just hiding it for now. I am planning to redo the PID controller later.
For PidControlledSystems, it used to be useful because of its utility
functions to wire up the old PID controller... That functionality still
works with the PR, and the whoever was using PidControlledSystems should
see no difference. But I intend to eliminate this class eventually or
replace by a ControlledSystem if there is a strong desire for it.

4. switched all kuka_iiwa_arm examples to the new Controller interface.

5. switched a test in simulator that had a pid controlled wired up by
hand.