Pull in the `fmt` string formatting library as a proper external dependency.... (#5607)
* Pull in the `fmt` string formatting library as a proper external dependency,
and have `maliput/utility` use it (for `GenerateObj()`, `GenerateUrdf()`).
Prior to this, this code was using the copy of the `fmt` library embedded in
`spdlog`.  However, `spdlog` is actually considered an optional dependency
(kind of), and stubs are provided if the real `spdlog` is not present (and
if bazel is lied-to in the right way).  This causes `maliput/utility` to
mysteriously fail to build if a user has asked for "spdlog stubs".  So...
now we will pull in `fmt` properly on its own.

Also, use `fmt::print()` uniformly throughout `GenerateObj()` --- `operator<<`
style output was still being used in some places --- places where `fmt` syntax
is more readable/terse.

Since `fmt` is the underlying formatting library used by our logging library,
it is generally useful to have that syntax consistently available for all of
our string formatting needs.

* Add `fmtlib/fmt` as a superbuild submodule.

* (In bazel build setup,) make `spdlog` dependent on the external `fmt` library
(and turn off the use of its embedded `fmt` library).

* Attempt to tell CMake to build `fmt` as an external library,
and have `spdlog` use it (instead of its internal copy).

* CMake build builds; and, reviewable comments addressed.