Correct demo to not overrun specified simulation duration (#5627)
To provide feedback on progress for very slow simulations, the demo would
take steps bounded by a "print time" interval so it could print progress
to the console.  However, if the simulation duration was *smaller* than
this print interval, the simulation would be extended to at least one
print duration.  The print duration is now constrained by the simulation
time.