Use tinyobjloader for loading .obj files
This pulls in tinyobjloader to do the .obj file loading.

There is a single caveat, it would be nice to have tinyobjloading do
the triangulation of non-triangular shapes, however it provides no
checks that such triangulations are valid, i.e. it assumes the user's
file is already correct. In addition, there is no way to check the
triangulations resulting from tinyobjloader's triangulation (you
have no idea what face was a triangle and what was not).

Subsequently, tinyobjloader here is used to load files which provides
several parsing benefits (e.g. ordering is no longer important), but
triangulation of complex shape checking is still handled manually.