Removed the superfluous l-value generator (#5749)
In solving another SFINAE problem, I realized that these tests are overly complex in their structure. We've introduced the GenerateLValue struct that isn't actually necessary; std::declval provides all the functionality we need.

1. In two cases (index and identifier tests) r-values are fine (declval<T>() has type T&&).
2. In the inertia tests, we do require l-values, becase we're doing in-place operations (e.g., *=). However, declval<T&>() has type T& && which collapses to T&. In short, the rvalue reference of an lvalue reference is an lvalue reference.

So, we simplify the voodoo using fewer arbitrary constructs and rely more directly on built-in C++ functionality.