Fixes #5745 ("Request to remove lane length restriction in DrawLaneArrow"). (#5768)
The "lane length restriction" is actually an assertion ensuring that
`DrawLaneArrow()` is not being asked to draw an arrow which pokes out beyond
the end of the lane.   The caller (`MarkLaneEnds()`) does in fact try to
calculate parameters which avoid that --- but apparently numeric precision
issues in double arithmetic can cause the assertion to be violated.

Solution:  nudge arrows inwards by `linear_tolerance()` of the RoadGeometry,
which provides plenty of buffer to avoid precision issues.

Also:  create a unit-test (based on @basicNew's test case) which exercises the failure.