ntroduces a copyable unique pointer and support logic.
The copyable_unique_pointer is a smart pointer that 
supports unique ownership semantics, but also has a copy 
constructor and assignment operator (when invoked, they 
copy the underlying object). The facilitates classes which use 
unique_ptr for ownership purposes to use a default copy 
constructor/assignment operator.

It also introduces the utility type_trait-like function: 
is_cloneable which can determine if a class is considered 
"cloneable" -- i.e., it has a const Clone() method that returns a 
unique_ptr of an assignable type.