Fix a bug in cmake_configure_file.py
The greedy portion of the regular expression could make
it so that lines never saw a carriage return.  This, in turn,
made it so that certain defines were never seen after generation.
To avoid this, make the regex have a + on the \r\n at the end,
to ensure that we always *must* have at least one newline
character.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>