Add PyCPS
Add PyCPS as a new Bazel external, along with its dependency, (the
Python module) semantic_version. This will allow us to generate
drake-config.cmake from drake.cps as part of the build process, rather
than doing it manually.

(This is not needed for the CMake build, since that generates
drake-config.cmake differently.)