Generate LCM's export header
Add a Bazel rule to generate export headers a la the functionality in
CMake (only supports ELF platforms for now). Use this to generate a real
header for LCM, which will allow us to turn ELF hidden visibility back
on for that library, and eventually export the library without needing
to muck about with exporting public compile definitions.