Start implementing installation in Bazel proper
Create Bazel rules to collect information on artifacts that should be
installed. Use these to declare installation rules for LCM.

This isn't complete; only the information gathering stage is
implemented. We still need to create a rule that generates an
installation script, and create a tool to perform the actual install
actions based on that script. However, just collecting the set of
artifacts to install and associating them with their full install
destination is significant progress.