Generate and install LCM exports
Update PyCPS, bringing in the ability to generate a version check file.
Add a script to generate a CPS for LCM (most of the content is
hard-coded, but the script extracts LCM's version information from
lcm_version.h). Implement generation of the remaining LCM artifacts.
Update LCM to pick up support for namespaced targets (which we get from
the CPS).

As of this change, it should be possible for CMake consumers to build
against the LCM that our new Bazel install logic installs.