Add ability to turn all assertion failures into exceptions
The exact semantics are per drake_assert.h comment changes, and
the toggle is in a new header drake_assert_toggle.h.  Note that
DRAKE_ABORT behavior is unchanged; only assertions are changed.

This also changes the thrown object (including DRAKE_THROW_UNLESS)
to be an exception subtype; this detail not exposed to the user in
any way, but might conceivably help diagnostics down the road, e.g.,
setting a breakpoint on its constructor.