Update package script to use new install framework
Remove various bits needed for the old installation mechanism. Rewrite
the packaging script to use the new install framework to create its
temporary install tree (which is then tar'd), rather than the old logic
that depended on a bunch of tar's produced by Bazel and manually
fiddling with the files.