Allow packages to reference dependencies
Add a file with common utilities for creating CPS's. Add ability for CPS
creation to depend on other CPS's. Add mechanisms to inject the version
information from another CPS into a CPS being created.

This allows a CPS to embed the exact version information of its
dependencies in a sane manner. For now, only spdlog does this. (Drake
proper will need to be updated eventually, but that can wait.)