Tweak handling of "foreign" install files
Modify public output_path to just return None for files outside the
current workspace, and to accept an optional override of the default
package root. Add an additional wrapper that is again private to
install.bzl that also takes a list of "allowed externals", which both
suppresses the warning about installing "foreign" files and can be used
to determine the correct partial path (using the new ability to override
the package root).

This improves the case of writing external install rules for a project
that is natively built by bazel but lacks its own install rules.