Add install rules for resource_tool
This changes targets destined for "bin" to use a package-qualified
path.  This is what Drake wants by default, and also happens to
work for almost all external packages because they have only a
single BUILD file.  (If some future package doesn't want this
behavior, we can add an install setting to alter it.)