Add renaming support to install framework
Add support for passing a map of install destinations to alternate file
names to the install framework. This allows files to be renamed on
install, which will be helpful for e.g. Java libraries, which Bazel
insists on prefixing with "lib", and which we often want to have named
the same as C++ libraries (which isn't otherwise possible because the
build labels must be unique).