Add utility to run pycodestyle on Bazel files
Create a new utility that wraps pycodestyle, disabling some checks and
adding another (to enforce space around named parameters). Create
bazel_lint.bzl to invoke this, adding the ability to ignore further
checks (and also adding this same ability to python_lint.bzl).

By default, bazel_lint disables some checks that are inimical to
readability, in addition to the checks which are hard-disabled by the
wrapper. (The latter are checks that are either impossible to enforce in
Skylark as they request use of non-existing features, or are contrary to
buildifier formatting.)

Note that this is not yet integrated into the build, as there are a
significant number of style issues currently present. This is being
merged as-is first to make the tool available, so that developers have a
way to check new work. We will start enforcing style later, as style
issues are fixed.